requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.103.2
fastapi-pagination==0.13.1
flake8==7.2.0
Flake8-pyproject==1.2.3
greenlet==3.2.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
priority==2.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.6
redis==5.2.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
taskgroup==0.2.2
tomli==2.2.1
types-pyasn1==0.6.0.20250208
types-python-jose==3.4.0.20250224
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
